Data.Heap.Double

Description

This module (and every module below it) is a duplicate of Data.Heap.Ord, specialized to keys of type Double.

Itself

data Heap a where

Spine-strict heap.

Bundled Patterns

pattern Empty :: Heap a
\(\mathcal{O}(1)\). Bidirectional pattern synonym matching an empty heap.

pattern (:<|) infixr 5
\(\mathcal{O}(1)\). Unidirectional pattern synonym analyzing the minimum of a non-empty heap.

Fields

:: Lookup a

-> Heap a

\(\mathcal{O}(\log n)\) thunk.

-> Heap a

Construct

empty :: Heap a

\(\mathcal{O}(1)\). Empty heap.

singleton :: Double -> a -> Heap a

\(\mathcal{O}(1)\). Heap with a single entry.

Single-key

Insert

insert :: Double -> a -> Heap a -> Heap a

\(\mathcal{O}(\log n)\). Insert a value in the heap with a given key.

Minimal-key

Lookup

lookupMin :: Heap a -> Maybe a

\(\mathcal{O}(1)\). Look up a value at the smallest key in the heap.

data Lookup a

Key together with the value.

Constructors

Lookup !Double a

lookupMinWithKey :: Heap a -> Maybe (Lookup a)

\(\mathcal{O}(1)\). Look up a value at the smallest key in the heap.

Map

adjustMin :: (a -> a) -> Heap a -> Heap a

\(\mathcal{O}(1)\). Apply a function to a value in the heap at the smallest key.

adjustMin' :: (a -> a) -> Heap a -> Heap a

\(\mathcal{O}(1)\). Apply a function to a value in the heap at the smallest key.

New value is evaluated to WHNF.

adjustMinWithKey :: (Double -> a -> a) -> Heap a -> Heap a

\(\mathcal{O}(1)\). Apply a function to a value in the heap at the smallest key.

adjustMinWithKey' :: (Double -> a -> a) -> Heap a -> Heap a

\(\mathcal{O}(1)\). Apply a function to a value in the heap at the smallest key.

New value is evaluated to WHNF.

Delete

deleteMin :: Heap a -> Heap a

\(\mathcal{O}(\log n)\). Delete a value at the smallest key in the heap.

Update

updateMin :: (a -> Maybe a) -> Heap a -> Heap a

\(\mathcal{O}(\log n)\). Update or delete a value in the heap at the smallest key.

The Maybe is evaluated to WHNF.

updateMinWithKey :: (Double -> a -> Maybe a) -> Heap a -> Heap a

\(\mathcal{O}(\log n)\). Update or delete a value in the heap at the smallest key.

The Maybe is evaluated to WHNF.

View

data ViewMin a

View of the minimum of the heap.

Constructors

(:<)

Fields

!(Lookup a)

(Heap a)

\(\mathcal{O}(\log n)\) thunk.

viewMin :: Heap a -> Maybe (ViewMin a)

\(\mathcal{O}(1)\). Analyze the minimum of the heap.

Full heap

Size

null :: Heap a -> Bool

\(\mathcal{O}(1)\). Check if the heap is empty.

size :: Heap a -> Word

\(\mathcal{O}(1)\). Calculate the number of entries stored in the heap.

Map

map :: (a -> b) -> Heap a -> Heap b

\(\mathcal{O}(n)\). Apply a function to every value in the heap.

map' :: (a -> b) -> Heap a -> Heap b

\(\mathcal{O}(n)\). Apply a function to every value in the heap.

New values are evaluated to WHNF.

mapWithKey :: (Double -> a -> b) -> Heap a -> Heap b

\(\mathcal{O}(n)\). Apply a function to every value in the heap.

mapWithKey' :: (Double -> a -> b) -> Heap a -> Heap b

\(\mathcal{O}(n)\). Apply a function to every value in the heap.

New values are evaluated to WHNF.

Fold

Left-to-right

foldl :: (b -> a -> b) -> b -> Heap a -> b

\(\mathcal{O}(n_R)\). Fold the heap left-to-right.

foldl' :: (b -> a -> b) -> b -> Heap a -> b

\(\mathcal{O}(n)\). Fold the heap left-to-right with a strict accumulator.

foldlWithKey :: (b -> Double -> a -> b) -> b -> Heap a -> b

\(\mathcal{O}(n_R)\). Fold the heap left-to-right.

foldlWithKey' :: (b -> Double -> a -> b) -> b -> Heap a -> b

\(\mathcal{O}(n)\). Fold the heap left-to-right with a strict accumulator.

Right-to-left

foldr :: (a -> b -> b) -> b -> Heap a -> b

\(\mathcal{O}(n_L)\). Fold the heap right-to-left.

foldr' :: (a -> b -> b) -> b -> Heap a -> b

\(\mathcal{O}(n)\). Fold the heap right-to-left with a strict accumulator.

foldrWithKey :: (Double -> a -> b -> b) -> b -> Heap a -> b

\(\mathcal{O}(n_L)\). Fold the heap right-to-left.

foldrWithKey' :: (Double -> a -> b -> b) -> b -> Heap a -> b

\(\mathcal{O}(n)\). Fold the heap right-to-left with a strict accumulator.

Monoid

foldMap :: Monoid m => (a -> m) -> Heap a -> m

\(\mathcal{O}(n_M)\). Map each value in a heap to a monoid and combine the results.

foldMapWithKey :: Monoid m => (Double -> a -> m) -> Heap a -> m

\(\mathcal{O}(n_M)\). Map each value in a heap to a monoid and combine the results.

Traverse

traverse :: Applicative f => (a -> f b) -> Heap a -> f (Heap b)

\(\mathcal{O}(n)\). Map each value in a heap to an action, evaluate these actions left-to-right and collect the results.

traverseWithKey :: Applicative f => (Double -> a -> f b) -> Heap a -> f (Heap b)

\(\mathcal{O}(n)\). Map each value in a heap to an action, evaluate these actions left-to-right and collect the results.

Union

union :: Heap a -> Heap a -> Heap a

\(\mathcal{O}(\log (\max (n_A, n_B))\). Union of two heaps.