Data.Heap.Ord

Description

Heap k a is a spine-strict binomial min-heap with a global root over keys of type (Ord k => k).

Binomial heaps have no notion of element order, as such:

Duplicate keys are permitted. Inserting a duplicate may put it anywhere on the heap, there is no guarantee a successive viewMin will pick the newer entry over the older one.

Arbitrary key access has \(\mathcal{O}(n)\) time complexity and operations that rely on it are therefore not provided.

Laziness

Evaluating the root of the heap (i.e. (_ :: Heap a)) to weak head normal form evaluates the entire spine of the heap to normal form.

Functions do not perform any additional evaluations unless their documentation directly specifies so.

Performance

Each function's time complexity is provided in the documentation.

\(n\) refers to the total number of entries in the heap. Parts of the heap are denoted using subscripts: \(n_L\) refers to the left side, \(n_R\) to the right side, and \(n_M\) to entries collected with the use of a Monoid.

Implementation

See the implementation section in Data.Heap.Ord.Unsafe for the explanation of the innerworkings.

Description of the binomial heap and the global root transformation can be found within the following papers:

Chris Okasaki, "Purely Functional Data Structures", September 1996, pages 68-72, https://www.cs.cmu.edu/~rwh/students/okasaki.pdf

Gerth Stølting Brodal and Chris Okasaki, "Optimal Purely Functional Priority Queues" , October 1996, pages 5-8 and 14-16, https://citeseerx.ist.psu.edu/document?doi=59776eb2358b4be265e7dae9df4f946dab25bf6f

Itself

data Heap k a where

Spine-strict heap.

Bundled Patterns

pattern Empty :: Heap k a
\(\mathcal{O}(1)\). Bidirectional pattern synonym matching an empty heap.

pattern (:<|) infixr 5
\(\mathcal{O}(1)\). Unidirectional pattern synonym analyzing the minimum of a non-empty heap.

Fields

:: Ord k

=> Lookup k a

-> Heap k a

\(\mathcal{O}(\log n)\) thunk.

-> Heap k a

Construct

empty :: Heap k a

\(\mathcal{O}(1)\). Empty heap.

singleton :: k -> a -> Heap k a

\(\mathcal{O}(1)\). Heap with a single entry.

Single-key

Insert

insert :: Ord k => k -> a -> Heap k a -> Heap k a

\(\mathcal{O}(\log n)\). Insert a value in the heap with a given key.

Minimal-key

Lookup

lookupMin :: Heap k a -> Maybe a

\(\mathcal{O}(1)\). Look up a value at the smallest key in the heap.

data Lookup k a

Key together with the value.

Constructors

Lookup !k a

lookupMinWithKey :: Heap k a -> Maybe (Lookup k a)

\(\mathcal{O}(1)\). Look up a value at the smallest key in the heap.

Map

adjustMin :: (a -> a) -> Heap k a -> Heap k a

\(\mathcal{O}(1)\). Apply a function to a value in the heap at the smallest key.

adjustMin' :: (a -> a) -> Heap k a -> Heap k a

\(\mathcal{O}(1)\). Apply a function to a value in the heap at the smallest key.

New value is evaluated to WHNF.

adjustMinWithKey :: (k -> a -> a) -> Heap k a -> Heap k a

\(\mathcal{O}(1)\). Apply a function to a value in the heap at the smallest key.

adjustMinWithKey' :: (k -> a -> a) -> Heap k a -> Heap k a

\(\mathcal{O}(1)\). Apply a function to a value in the heap at the smallest key.

New value is evaluated to WHNF.

Delete

deleteMin :: Ord k => Heap k a -> Heap k a

\(\mathcal{O}(\log n)\). Delete a value at the smallest key in the heap.

Update

updateMin :: Ord k => (a -> Maybe a) -> Heap k a -> Heap k a

\(\mathcal{O}(\log n)\). Update or delete a value in the heap at the smallest key.

The Maybe is evaluated to WHNF.

updateMinWithKey :: Ord k => (k -> a -> Maybe a) -> Heap k a -> Heap k a

\(\mathcal{O}(\log n)\). Update or delete a value in the heap at the smallest key.

The Maybe is evaluated to WHNF.

View

data ViewMin k a

View of the minimum of the heap.

Constructors

(:<)

Fields

!(Lookup k a)

(Heap k a)

\(\mathcal{O}(\log n)\) thunk.

viewMin :: Ord k => Heap k a -> Maybe (ViewMin k a)

\(\mathcal{O}(1)\). Analyze the minimum of the heap.

Full heap

Size

null :: Heap k a -> Bool

\(\mathcal{O}(1)\). Check if the heap is empty.

size :: Heap k a -> Word

\(\mathcal{O}(1)\). Calculate the number of entries stored in the heap.

Map

map :: (a -> b) -> Heap k a -> Heap k b

\(\mathcal{O}(n)\). Apply a function to every value in the heap.

map' :: (a -> b) -> Heap k a -> Heap k b

\(\mathcal{O}(n)\). Apply a function to every value in the heap.

New values are evaluated to WHNF.

mapWithKey :: (k -> a -> b) -> Heap k a -> Heap k b

\(\mathcal{O}(n)\). Apply a function to every value in the heap.

mapWithKey' :: (k -> a -> b) -> Heap k a -> Heap k b

\(\mathcal{O}(n)\). Apply a function to every value in the heap.

New values are evaluated to WHNF.

Fold

Left-to-right

foldl :: (b -> a -> b) -> b -> Heap k a -> b

\(\mathcal{O}(n_R)\). Fold the heap left-to-right.

foldl' :: (b -> a -> b) -> b -> Heap k a -> b

\(\mathcal{O}(n)\). Fold the heap left-to-right with a strict accumulator.

foldlWithKey :: (b -> k -> a -> b) -> b -> Heap k a -> b

\(\mathcal{O}(n_R)\). Fold the heap left-to-right.

foldlWithKey' :: (b -> k -> a -> b) -> b -> Heap k a -> b

\(\mathcal{O}(n)\). Fold the heap left-to-right with a strict accumulator.

Right-to-left

foldr :: (a -> b -> b) -> b -> Heap k a -> b

\(\mathcal{O}(n_L)\). Fold the heap right-to-left.

foldr' :: (a -> b -> b) -> b -> Heap k a -> b

\(\mathcal{O}(n)\). Fold the heap right-to-left with a strict accumulator.

foldrWithKey :: (k -> a -> b -> b) -> b -> Heap k a -> b

\(\mathcal{O}(n_L)\). Fold the heap right-to-left.

foldrWithKey' :: (k -> a -> b -> b) -> b -> Heap k a -> b

\(\mathcal{O}(n)\). Fold the heap right-to-left with a strict accumulator.

Monoid

foldMap :: Monoid m => (a -> m) -> Heap k a -> m

\(\mathcal{O}(n_M)\). Map each value in a heap to a monoid and combine the results.

foldMapWithKey :: Monoid m => (k -> a -> m) -> Heap k a -> m

\(\mathcal{O}(n_M)\). Map each value in a heap to a monoid and combine the results.

Traverse

traverse :: Applicative f => (a -> f b) -> Heap k a -> f (Heap k b)

\(\mathcal{O}(n)\). Map each value in a heap to an action, evaluate these actions left-to-right and collect the results.

traverseWithKey :: Applicative f => (k -> a -> f b) -> Heap k a -> f (Heap k b)

\(\mathcal{O}(n)\). Map each value in a heap to an action, evaluate these actions left-to-right and collect the results.

Union

union :: Ord k => Heap k a -> Heap k a -> Heap k a

\(\mathcal{O}(\log (\max (n_A, n_B))\). Union of two heaps.