Data.Heap.Ord.Unsafe

Description

Data structure internals, helper operations and unsafe functions.

Implementation

The implementation mostly copies the heap described by Okasaki (and on the Wikipedia page). There are two differences:

Ranks are not stored in nodes; for the top-level forest they're instead inferred from heap's size. All other ranks are redundant information, as pointed out by Okasaki on page 70 of his "Purely Functional Data Structures" paper.

The heap has a global root for \(\mathcal{O}(1)\) minimal-entry access. This transformation does not cost anything as the heap already has to hold its size at the root to satisfy the first point.

Operations are therefore altered appropriately:

insert

The number of top-level trees touched by the insertion is the same as the number of consecutive lowermost 1 bits in the size of the tree.

meld

Ranks and whichever heap will bottom out first is known beforehand from the sizes. Melding continues until the smaller heap contains no more ranks, the remaining carry bit (if any) can then be inserted into the remainder of the larger heap.

delete-min

Given a heap \(A\), deletion of the node with the minimal key leaves a forest \(B\) which needs to be reversed and melded with the remaining heap. \(B\) is guaranteed to have the size of \(2^k - 1\) where \(k \ge 1\). \(A\) is known to be larger than \(B\). Any carry during the meld results in repeated carries for the rest of the operation. The deleted node is guaranteed to be the lowermost node in what remains of \(A\) after the meld; the remaining carry bit (if any) will have the same rank as the deleted node.

Documentation

type Size = Word

Number of entries in the heap, excluding the entry with the smallest key.

data Heap k a

Spine-strict heap.

Constructors

Heap

Fields

!Size

!k

Smallest key in the heap.

a

Value under the smallest key.

!(Forest k a)

None

data Forest k a

A forest is a cons-list of binomial trees, such that:

rank progression on each level is strictly monotonic;

keys on subsequent levels are greater than or equal to the key of the node above.

The top-level forest has its ranks in ascending order; the ranks correspond to the base-2 powers in the Size of the heap.

Forests on every other level have their ranks in descending order; the ranks are packed densely. As a consequence, for any non-top-level tree of rank \(k\), except for \(k = 0\), the next tree and the largest tree below are both of rank \(k - 1\).

Constructors

Tree

Fields

!k

a

!(Forest k a)

Forest below this node.

!(Forest k a)

Next tree in the forest.

Nil