Data.Heap.Word.Debug

Description

Safe functions for datatype introspection.

Show

showsHeap :: (a -> ShowS) -> Heap a -> ShowS

\(\mathcal{O}(n)\). Shows the internal structure of the heap.

Eq

eqHeap :: (a -> b -> Bool) -> Heap a -> Heap b -> Bool

\(\mathcal{O}(n)\). Checks whether the internal representations of two heaps are equal.

Validate

data Validity

Whether the heap is well-formed.

Constructors

Valid

Invalid Reason

data Reason

Reason for why the heap is considered malformed.

Constructors

KeyBelow !Word
Forest below the node with this key holds a smaller key.

MalformedRank !Word
Rank below the node with this key is malformed.

NotMonotone ![Int]
Stored ranks are not monotonically ascending.

RankMismatch !Size ![Int]
Stored ranks do not match the size.

validate :: Heap a -> Validity

\(\mathcal{O}(n)\). Checks whether the heap is well-formed.