Data.Heap.Word.Unsafe

Description

Data structure internals, helper operations and unsafe functions.

Documentation

type Size = Word

Number of entries in the heap, excluding the entry with the smallest key.

data Heap a

Spine-strict heap.

Constructors

Heap

Fields

!Size

!Word

Smallest key in the heap.

a

Value under the smallest key.

!(Forest a)

None

data Forest a

A forest is a cons-list of binomial trees, such that:

rank progression on each level is strictly monotonic;

keys on subsequent levels are greater than or equal to the key of the node above.

The top-level forest has its ranks in ascending order; the ranks correspond to the base-2 powers in the Size of the heap.

Forests on every other level have their ranks in descending order; the ranks are packed densely. As a consequence, for any non-top-level tree of rank \(k\), except for \(k = 0\), the next tree and the largest tree below are both of rank \(k - 1\).

Constructors

Tree

Fields

!Word

a

!(Forest a)

Forest below this node.

!(Forest a)

Next tree in the forest.

Nil